pin-service/api/requirements.txt
fastapi
uvicorn
sqlalchemy
geoalchemy2
shapely
geojson
psycopg2-binary
requests
pytest
httpx
python-multipart
pyjwt
python-jose[cryptography]
passlib[bcrypt]